requirements.txt
playwright>=1.46
beautifulsoup4>=4.12
lxml>=5.2
pydantic>=2.9
python-slugify>=8.0
tenacity>=9.0
python-dotenv>=1.0
pytest>=8.0

